Heya

+++ push_swap.h
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   push_swap.h                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: kbarbry <kbarbry@student.42lyon.fr>        +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/11/27 17:59:09 by kbarbry           #+#    #+#             */
+/*   Updated: 2021/11/27 18:14:41 by kbarbry          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef PUSH_SWAP_H
 # define PUSH_SWAP_H
 
@@ -7,11 +19,11 @@
 
 //utilities.c
 
-int		ft_atoi(char *str);
-char	*ft_calloc(int size);
+long int	ft_atoi(char *str);
+char	    *ft_calloc(int size);
 
 //error.c
 
-int		ft_error(void);
+int		    ft_error(void);
 
 #endif
